refac: Separate transport from channel and split it to uni-di & bi-di
